generalized scripting interfaces
